Factor out excess logic from views
